move home screen state classes
