+++ app/src/main/java/com/example/tudeeapp/presentation/screen/home/HomeUiState.kt
@@ -1,4 +1,4 @@
-package com.example.tudeeapp.presentation.screen.home.state
+package com.example.tudeeapp.presentation.screen.home
 
 
 data class HomeUiState(
