+++ app/src/main/java/com/example/tudeeapp/presentation/screen/home/HomeViewModel.kt
@@ -9,8 +9,6 @@ import com.example.tudeeapp.domain.exception.TudeeException
 import com.example.tudeeapp.domain.models.Task
 import com.example.tudeeapp.domain.models.TaskPriority
 import com.example.tudeeapp.domain.models.TaskStatus
-import com.example.tudeeapp.presentation.screen.home.state.HomeUiState
-import com.example.tudeeapp.presentation.screen.home.state.TaskUiState
 import com.example.tudeeapp.presentation.screen.home.utils.getToday
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.flow.MutableStateFlow
