+++ app/src/main/java/com/example/tudeeapp/presentation/screen/home/SliderState.kt
@@ -1,4 +1,4 @@
-package com.example.tudeeapp.presentation.screen.home.state
+package com.example.tudeeapp.presentation.screen.home
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
